Add OMRPortLib to Jit test compiler environment
Jit test harness initialize a OMRPortLibrary but never add it to the
compiler environment. Some platforms like z rely on port library for
some operations like getting the cpu features. As a result, z always
defaults on the minimum supported version to get cpu features in
evaluators while running compiler tests which does not match the actual
cpu and cause false failures.
This change add the OMRPortLibrary to the test compiler environment.

signed-off-by: Ehsan Kiani Far <ehsan.kianifar@gmail.com>